refactor(opencode): extract trade memory core modules
